backend/requirements.txt
flask
flask-cors
numpy
pandas
torch
scikit-learn
supabase
python-dotenv
requests
matplotlib
seaborn
mediapipe
opencv-python
PyJWT[crypto]